requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
cryptography>=3.4.0
requests>=2.28.0
websockets>=10.0
web3>=6.0.0
eth-account>=0.8.0
hyperliquid-python-sdk>=0.2.6
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
itsdangerous>=2.1.0
psutil>=5.9.0
psycopg2-binary>=2.9.9